backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
langchain-groq==0.1.6
email-validator==2.1.0